+++ apina-core/src/main/kotlin/fi/evident/apina/model/settings/BrandedPrimitiveType.kt
@@ -1,0 +1,6 @@
+package fi.evident.apina.model.settings
+
+import fi.evident.apina.model.type.ApiType
+import fi.evident.apina.model.type.ApiTypeName
+
+class BrandedPrimitiveType(val brandedType: ApiTypeName, val implementationType: ApiType)